requirements.txt
django
mysqlclient
djangorestframework
pillow